fix:map text
