+++ apps/rahat-ui/src/sections/projects/sms-voucher/const.ts
@@ -1,21 +1,10 @@
-export function mapStatus(status: string) {
-  const statusMapping: any = {
-    CHECKED: 'Eye Checkup',
-    PURCHASE_OF_GLASSES: 'Purchase of Glasses',
-    READING_GLASSES: 'Reading Glasses',
-    SUN_GLASSES: 'Sun Glasses',
-    PRESCRIBED_LENSES: 'Prescribed Lenses',
-    REDEEMED: 'Redeemed',
-    NOT_REDEEMED: 'Not Redeemed',
-    yes: 'Yes',
-    no: 'No',
-    skip: 'Skip',
-  };
-  return statusMapping[status] || '-';
-}
 export function mapTopic(topic: string) {
   const topicMapping: any = {
     'Walk In Beneficiary Added': 'Voucher Redeemed',
+    NOT_REDEEM_STATS: 'Total Inactive Consumers',
+    NOT_REQUIRED: 'Not Redeemed',
+    REDEMPTION_STATS: 'Glass Purchase Type',
+    VOUCHER_USAGE_TYPE_STATS: 'Voucher Usage Type',
   };
   return topicMapping[topic] || topic;
 }
