+++ libs/query/src/lib/el-kenya/index.ts
@@ -1,3 +1,4 @@
 export * from './subgraph';
 export * from './contracts/hooks';
 export * from './project-actions';
+export * from './utils/index';
