+++ libs/query/src/lib/el-kenya/utils/index.ts
@@ -17,3 +17,19 @@ export const formatTransaction = (trans: any) => ({
 
 export const mapTransactions = (transactions: any[]) =>
   transactions.map(formatTransaction);
+
+export function mapStatus(status: string) {
+  const statusMapping: any = {
+    CHECKED: 'Eye Checkup',
+    PURCHASE_OF_GLASSES: 'Purchase of Glasses',
+    READING_GLASSES: 'Reading Glasses',
+    SUN_GLASSES: 'Sun Glasses',
+    PRESCRIBED_LENSES: 'Prescribed Lenses',
+    REDEEMED: 'Redeemed',
+    NOT_REDEEMED: 'Not Redeemed',
+    yes: 'Yes',
+    no: 'No',
+    skip: 'Skip',
+  };
+  return statusMapping[status] || '-';
+}
